--- docs/EXhallsStrapi.json
@@ -1,64 +1,0 @@
-[
-    {
-        "id": 2,
-        "Name": null,
-        "Number": null,
-        "NumberOfSeats": null,
-        "cinema": null,
-        "created_by": {
-            "id": 3,
-            "firstname": "Liis",
-            "lastname": "Käsper",
-            "username": null
-        },
-        "updated_by": {
-            "id": 3,
-            "firstname": "Liis",
-            "lastname": "Käsper",
-            "username": null
-        },
-        "created_at": "2020-08-26T13:23:11.303Z",
-        "updated_at": "2020-08-26T13:23:11.328Z",
-        "Floor": null,
-        "Name_et": null,
-        "Name_en": null,
-        "Name_ru": null,
-        "name_et": "Test88",
-        "number": 88,
-        "numberOfSeats": 20,
-        "floor": "Second",
-        "name_en": "Test88",
-        "name_ru": "Test88"
-    },
-    {
-        "id": 3,
-        "Name": null,
-        "Number": null,
-        "NumberOfSeats": null,
-        "cinema": null,
-        "created_by": {
-            "id": 3,
-            "firstname": "Liis",
-            "lastname": "Käsper",
-            "username": null
-        },
-        "updated_by": {
-            "id": 3,
-            "firstname": "Liis",
-            "lastname": "Käsper",
-            "username": null
-        },
-        "created_at": "2020-08-26T13:39:38.752Z",
-        "updated_at": "2020-08-26T13:39:38.771Z",
-        "Floor": null,
-        "Name_et": null,
-        "Name_en": null,
-        "Name_ru": null,
-        "name_et": "Test8",
-        "number": null,
-        "numberOfSeats": null,
-        "floor": null,
-        "name_en": null,
-        "name_ru": null
-    }
-]
