--- docs/EXorgStrapi.json
@@ -1,38 +1,0 @@
-{
-    "id": 3,
-    "Name": null,
-    "Description": null,
-    "created_by": {
-        "id": 3,
-        "firstname": "Liis",
-        "lastname": "Käsper",
-        "username": null
-    },
-    "updated_by": {
-        "id": 3,
-        "firstname": "Liis",
-        "lastname": "Käsper",
-        "username": null
-    },
-    "created_at": "2020-08-26T09:52:26.012Z",
-    "updated_at": "2020-08-26T13:52:14.433Z",
-    "Name_et": null,
-    "Name_en": null,
-    "Name_ru": null,
-    "Name_original": null,
-    "Type": null,
-    "name": "test",
-    "type": "ProductionCompany",
-    "gender": null,
-    "person": null,
-    "nameTranslated": [
-        {
-            "id": 3547,
-            "et": "testname",
-            "en": null,
-            "ru": null
-        }
-    ],
-    "description": null,
-    "logo": null
-}
